requirements-dev.txt
black==24.8.0
flake8==7.1.1
flake8-bugbear==24.8.19
pytest==7.4.3
pytest-asyncio==0.23.7
pre-commit==3.8.0
Pillow>=10.0
